[Refactor] vite.config.ts (#392)
