+++ utils/vite.ts
@@ -1,0 +1,31 @@
+import { type PluginOption } from 'vite';
+import makeManifest from './plugins/make-manifest';
+import customDynamicImport from './plugins/custom-dynamic-import';
+import addHmr from './plugins/add-hmr';
+import watchRebuild from './plugins/watch-rebuild';
+import inlineVitePreloadScript from './plugins/inline-vite-preload-script';
+
+export const getPlugins = (isDev: boolean): PluginOption[] => [
+  makeManifest({ getCacheInvalidationKey }),
+  customDynamicImport(),
+  // You can toggle enable HMR in background script or view
+  addHmr({ background: true, view: true, isDev }),
+  isDev && watchRebuild({ afterWriteBundle: regenerateCacheInvalidationKey }),
+  // For fix issue#177 (https://github.com/Jonghakseo/chrome-extension-boilerplate-react-vite/issues/177)
+  inlineVitePreloadScript(),
+];
+
+const cacheInvalidationKeyRef = { current: generateKey() };
+
+export function getCacheInvalidationKey() {
+  return cacheInvalidationKeyRef.current;
+}
+
+function regenerateCacheInvalidationKey() {
+  cacheInvalidationKeyRef.current = generateKey();
+  return cacheInvalidationKeyRef;
+}
+
+function generateKey(): string {
+  return `${Date.now().toFixed()}`;
+}
