+++ vite.config.ts
@@ -2,48 +2,28 @@
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
 import path, { resolve } from 'path';
-import makeManifest from './utils/plugins/make-manifest';
-import customDynamicImport from './utils/plugins/custom-dynamic-import';
-import addHmr from './utils/plugins/add-hmr';
-import watchRebuild from './utils/plugins/watch-rebuild';
-import inlineVitePreloadScript from './utils/plugins/inline-vite-preload-script';
+import { getCacheInvalidationKey, getPlugins } from './utils/vite';
 
 const rootDir = resolve(__dirname);
 const srcDir = resolve(rootDir, 'src');
 const pagesDir = resolve(srcDir, 'pages');
-const assetsDir = resolve(srcDir, 'assets');
-const outDir = resolve(rootDir, 'dist');
-const publicDir = resolve(rootDir, 'public');
 
 const isDev = process.env.__DEV__ === 'true';
 const isProduction = !isDev;
-
-// ENABLE HMR IN BACKGROUND SCRIPT
-const enableHmrInBackgroundScript = true;
-const cacheInvalidationKeyRef = { current: generateKey() };
 
 export default defineConfig({
   resolve: {
     alias: {
       '@root': rootDir,
       '@src': srcDir,
-      '@assets': assetsDir,
+      '@assets': resolve(srcDir, 'assets'),
       '@pages': pagesDir,
     },
   },
-  plugins: [
-    makeManifest({
-      getCacheInvalidationKey,
-    }),
-    react(),
-    customDynamicImport(),
-    addHmr({ background: enableHmrInBackgroundScript, view: true }),
-    isDev && watchRebuild({ afterWriteBundle: regenerateCacheInvalidationKey }),
-    inlineVitePreloadScript(),
-  ],
-  publicDir,
+  plugins: [...getPlugins(isDev), react()],
+  publicDir: resolve(rootDir, 'public'),
   build: {
-    outDir,
+    outDir: resolve(rootDir, 'dist'),
     /** Can slow down build speed. */
     // sourcemap: isDev,
     minify: isProduction,
@@ -81,15 +61,3 @@ export default defineConfig({
     setupFiles: './test-utils/vitest.setup.js',
   },
 });
-
-function getCacheInvalidationKey() {
-  return cacheInvalidationKeyRef.current;
-}
-function regenerateCacheInvalidationKey() {
-  cacheInvalidationKeyRef.current = generateKey();
-  return cacheInvalidationKeyRef;
-}
-
-function generateKey(): string {
-  return `${Date.now().toFixed()}`;
-}
